backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
google-generativeai==0.7.0
python-dotenv==1.0.1
pydantic[email]==2.7.1
